fix(scripts): adapt scripts so we can expose them thru poetry.
Moved the scripts within the package and renamed them with the right
Python extension. They could not be imported properly before since they
were lacking the .py extension and did not define a main function.

Tested with `poetry build` and `pip install` on the generated wheel
locally against test files from unblob.
