+++ pyproject.toml
@@ -34,9 +34,9 @@ exclude = []
 exclude = ["build"]
 
 [tool.poetry.scripts]
-ubireader_display_blocks = "scripts.ubireader_display_blocks:main"
-ubireader_display_info = "scripts.ubireader_display_info:main"
-ubireader_extract_files = "scripts.ubireader_extract_files:main"
-ubireader_extract_images = "scripts.ubireader_extract_images:main"
-ubireader_list_files = "scripts.ubireader_list_files:main"
-ubireader_utils_info = "scripts.ubireader_utils_info:main"
+ubireader_display_blocks = "ubireader.scripts.ubireader_display_blocks:main"
+ubireader_display_info = "ubireader.scripts.ubireader_display_info:main"
+ubireader_extract_files = "ubireader.scripts.ubireader_extract_files:main"
+ubireader_extract_images = "ubireader.scripts.ubireader_extract_images:main"
+ubireader_list_files = "ubireader.scripts.ubireader_list_files:main"
+ubireader_utils_info = "ubireader.scripts.ubireader_utils_info:main"
