+++ ubireader/scripts/__init__.py
@@ -1,0 +1,1 @@
+
